src/main/java/com/example/demo/Product/queryhandlers/GetProductQueryHandler.java
package com.example.demo.Product.queryhandlers;

import com.example.demo.Exceptions.ProductNotFoundException;
import com.example.demo.Product.Model.Product;
import com.example.demo.Product.Model.ProductDTO;
import com.example.demo.Product.ProductRepository;
import com.example.demo.Query;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.stereotype.Service;

import java.util.Optional;

@Service
public class GetProductQueryHandler implements Query<Integer, ProductDTO> {
    @Autowired
    ProductRepository productRepository;
    @Override
    public ResponseEntity<ProductDTO> execute(Integer id){
        Optional<Product> product= productRepository.findById(id);
        if(product.isEmpty()){
            throw new ProductNotFoundException();
        }
        ProductDTO productDTO = new ProductDTO(product.get());
        return ResponseEntity.ok(productDTO);
    }

}
